moved application packages tests into their own cs file
