+++ src/ResourceManagement/Batch/Batch.Tests/Batch.Tests.csproj
@@ -10,6 +10,7 @@
   <ItemGroup>
     <Compile Include="Helpers\RecordedDelegatingHandler.cs" />
     <Compile Include="InMemoryTests\AccountTests.InMemory.cs" />
+    <Compile Include="InMemoryTests\ApplicationPackageTests.InMemory.cs" />
     <Compile Include="InMemoryTests\SubscriptionTests.InMemory.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SpecTestSupport.Batch\BatchManagementClientUndoHandlers.cs" />
